Move Regexes and Colour class to separate files
